Clean up server state on shutdown.
This adds a Stop() method to Session-s and Handler-s.
This is called by ServeConn when its recv/send loop [serve()]
returns.  Also, serve() now cancels all outstanding tags
when the connection is terminated.
