+++ context.go
@@ -2,6 +2,7 @@ package p9p
 
 import (
 	"context"
+	"time"
 )
 
 type contextKey string
@@ -23,4 +24,25 @@ func GetVersion(ctx context.Context) string {
 		return ""
 	}
 	return v
+}
+
+// Simple context representing a past-due deadline.
+type CancelledCtxt struct{}
+
+func (_ CancelledCtxt) Deadline() (time.Time, bool) {
+    return time.Time{}, false
+}
+
+func (_ CancelledCtxt) Done() <-chan struct{} {
+     ch := make(chan struct{})
+     close(ch)
+     return ch
+}
+
+func (_ CancelledCtxt) Err() error {
+    return context.Canceled
+}
+
+func (_ CancelledCtxt) Value(key any) (val any) {
+    return nil
 }
