+++ session.go
@@ -39,4 +39,7 @@ type Session interface {
 	// can be affected by negotiating or the level of support provided by the
 	// session implementation.
 	Version() (msize int, version string)
+
+    // Called at server shutdown.
+    Stop(error) error
 }
